Remove X-Powered-By token
It is caught by the default header-field now and has no directives
that I know of. Ideally all tokens in Extensions would add directives
*if* the role is mixed into the grammar, but making this work while
still working with Actions will require more understanding. But if
Extensions is made optional we could put all the `known-header`
tokens there as well. This will make it so if the grammar
`does Extensions` it will apply header-specific rules and captures,
otherwise it will return basic `Key: Value` pairs.